fixed requestId type issues addressed in the comments
the solution that suggests that i make upload media as a nextfunction to updaterequest or create request didn't work in my previous tests so i had to settle to making them seperate , if any of you has an idea on how to fix it , or can see if you can do it in a normal (not using postman) request , let me know and it'll be fixed

+++ Back-End/controller/jobRequestController/deleteRequest.ts
@@ -4,8 +4,15 @@ import { Request , Response } from "express";
 const deleteRequest = async (req : Request , res : Response) => {
     try {
         const requestId : number= +req.params.requestId;
+        if (Number.isNaN(requestId)){
+          res.status(400).json({
+          message: 'Please provide request id',
+          success : false
+          });
+          return;
+          }
         await pool.query(`DELETE FROM request WHERE id=$1`,[requestId]);
-        res.status(201).json({ 
+        res.status(200).json({ 
             message : "request deleted successfully",
             success : true
         })
